refactor listView adapter e metodi annesi per i diversi fragment che usavano lo stesso codice duplicato
