+++ app/src/main/java/it/unimib/brewday/ui/gestisci_ingredienti/IngredientiFragment.java
@@ -36,10 +36,6 @@ public class IngredientiFragment extends Fragment {
 
     List<Ingrediente> listaIngredienti;
 
-    int posizionePrecedente = -1;
-    EditText quantitaIngredientePrecedente;
-
-    Ingrediente ingredientePrecedente;
 
 
     public IngredientiFragment() {
@@ -80,28 +76,18 @@ public class IngredientiFragment extends Fragment {
 
                 adapterListViewListaIngredientiDisponibili = new AdapterListViewListaIngredientiDisponibili(getContext(), 0, listaIngredienti, R.layout.lista_ingredienti_singoli, new AdapterListViewListaIngredientiDisponibili.OnItemClickListener() {
                     @Override
-                    public void onAddIngredienteClick(Ingrediente ingrediente, int position, EditText quantitaIngrediente) {
-
-                        aggiungiQuantitaIngrediente(ingrediente, position, quantitaIngrediente);
-                        aggiornaDBIngrediente( verificaIngrediente(ingrediente, quantitaIngrediente));
+                    public void onAddIngredienteClick(Ingrediente ingrediente) {
+                        aggiornaDBIngrediente(ingrediente);
                     }
 
                     @Override
-                    public void onRemoveIngredienteClick(Ingrediente ingrediente, int position, EditText quantitaIngrediente) {
-
-                        if (ingrediente.getQuantitaPosseduta() < 1) {
-                            Snackbar.make(view, "Non si possono avere ingredienti negativi", LENGTH_SHORT).show();
-                        }else {
-                            togliQuantitaIngrediente(ingrediente, position, quantitaIngrediente);
-                            aggiornaDBIngrediente( verificaIngrediente(ingrediente, quantitaIngrediente));                     }
-
+                    public void onRemoveIngredienteClick(Ingrediente ingrediente) {
+                        aggiornaDBIngrediente(ingrediente);
                     }
-                }, (ingrediente, quantitaIngrediente, position) -> {
-                    resetQuantitaLasciatoTestoVuoto(ingrediente, quantitaIngrediente);
-                    inizializzaPositionePrecedente(ingrediente, position, quantitaIngrediente);
-                    controlloCambioSelezione(ingrediente, position, quantitaIngrediente);
-                    rispostaInvioTastiera(ingrediente, position, quantitaIngrediente);
-                }, true);
+                }, (ingrediente) ->
+
+                    aggiornaDBIngrediente( ingrediente)
+                , true);
 
                 listViewIngredientiDispobili.setAdapter(adapterListViewListaIngredientiDisponibili);
                 listViewIngredientiDispobili.setDivider(null);
@@ -112,85 +98,12 @@ public class IngredientiFragment extends Fragment {
         });
     }
 
-    public Ingrediente verificaIngrediente(Ingrediente ingrediente, EditText quantitaIngrediente) {
-
-        if (quantitaIngrediente.getText().length() == 0) {
-            ingrediente.setQuantitaPosseduta(0);
-            quantitaIngrediente.setText("0");
-        } else {
-            quantitaIngrediente.setText(String.valueOf(Integer.parseInt(quantitaIngrediente.getText().toString())));
-            ingrediente.setQuantitaPosseduta(Integer.parseInt(quantitaIngrediente.getText().toString()));
-        }
-
-        return ingrediente;
-    }
-
-    public int quantitaBottone(int position) {
-        if (position == 0) {
-            return 1;
-        } else {
-            return 10;
-        }
-
-
-    }
-    public void inizializzaPositionePrecedente(Ingrediente ingrediente, int position, EditText quantitaIngrediente) {
-
-        if (posizionePrecedente == -1) {
-            posizionePrecedente = position;
-            quantitaIngredientePrecedente = quantitaIngrediente;
-            ingredientePrecedente = ingrediente;
-        }
-
-    }
-    public void controlloCambioSelezione(Ingrediente ingrediente, int position, EditText quantitaIngrediente) {
-
-        if (posizionePrecedente != position) {
-            aggiornaDBIngrediente(verificaIngrediente(ingredientePrecedente, quantitaIngredientePrecedente));
-            posizionePrecedente = position;
-            quantitaIngredientePrecedente = quantitaIngrediente;
-            ingredientePrecedente = ingrediente;
-        }
-
-    }
-
-
-    public void togliQuantitaIngrediente(Ingrediente ingrediente, int position, EditText quantitaIngrediente){
-        if (ingrediente.getQuantitaPosseduta() < 10 && quantitaBottone(position) == 10) {
-            ingrediente.setQuantitaPosseduta(0);
-        } else {
-            ingrediente.setQuantitaPosseduta(verificaIngrediente(ingrediente ,quantitaIngrediente).getQuantitaPosseduta() - quantitaBottone(position));
-        }
-        quantitaIngrediente.setText(ingrediente.getQuantitaAssolutaToString());
-
-    }
-
-    public void aggiungiQuantitaIngrediente(Ingrediente ingrediente, int position, EditText quantitaIngrediente){
-            ingrediente.setQuantitaPosseduta(verificaIngrediente(ingrediente, quantitaIngrediente).getQuantitaPosseduta() + quantitaBottone(position));
-            quantitaIngrediente.setText(ingrediente.getQuantitaAssolutaToString());
-
-    }
-
     public void aggiornaDBIngrediente(Ingrediente ingrediente){
         ingredienteViewModel.updateIngrediente(ingrediente);
 
 
     }
 
-    public void resetQuantitaLasciatoTestoVuoto(Ingrediente ingrediente, EditText quantitaIngrediente) {
-        if (quantitaIngrediente.getText().length() == 0) {
-            ingrediente.setQuantitaPosseduta(0);
-            aggiornaDBIngrediente(ingrediente);
-        }
-    }
 
-    public void rispostaInvioTastiera(Ingrediente ingrediente, int position, EditText quantitaIngrediente){
-        quantitaIngrediente.setOnKeyListener((v, keyCode, event) -> {
-            if (event.getAction() == KeyEvent.ACTION_DOWN && keyCode == KeyEvent.KEYCODE_ENTER) {
-                aggiornaDBIngrediente( verificaIngrediente(ingrediente, quantitaIngrediente));
-            }
-            return false;
-        });
-    }
 
 }
